Stationarity Test
Hi Guys, 

I managed, also thanks to Marcus' suggerstion, the stationarity tests for our entire panel. 

Thanks to using the apply command, I was able to use one loop less (the one looping over the variables). It is a shame that I was not able avoid looping over the companies. I am a bit of a squarehead in that matter. 

Please check whether I apply the ADF-Test in the right fashion. Maybe we could apply another test as well - yet only if this really helps us to capture other types of non-stationarity. 

Regards
Tim