Pre 2012 Stationarity tests added
Hi guys, 

it seems that most of the non-stationarity in Net Income indeed disappears excluding the years 2013-2015, i.e. the most interesting ones. 

I therefore might conclude that the panel data model the guys set up is nice - yet since it cannot deal with nonstationary variables (Net Income in this case) it might not be the adequate one for analyzing what happened in 2013 - 2015 :(