Create graphs with no loop s
this is the folder in which I have stored the graphs with no loops on 16.06.2016

Still, I would actually stand by using loops at least for the plots of stock prices, Assets/Mcap ... in absoltute terms, because we can keep the time axis and the y axis labels. 

For the histograms, this does not matter so much, yet I hold this to be more informative, too. 

Apart from that, if we loop by enterprise, we can parmit ourselves to provide the plots with a larger margin - which means that we can increase the font size of the labels.